Renamed history to news; changed sort
