+++ components/rockets/SpecItem.tsx
