+++ pages/rockets/[rocket].tsx
